Fix yate hangup handling
The hangup handling for IVRs must have always been broken as the
chan.hangup message seems to never reach an IVR. When a sip client
hangs-up, yate internally creates a chan.hangup message and also
enqueues it for engine processing but before it is returning to
the event loop, it calls disconnect() on the channel. This causes
the extmodule channel to be disconnected and its refcount is lowered.
As a consequence, the destructor is called even before the chan.hangup
message had a chance of being processed. The destructor closes stdin
of the extmodule and sends the process SIGTERM if it isn't super quick
in shutting down on its own.

Consequently, the IVR code now reacts on the message channel being closed
remotely and uses this to trigger the hangup event. In order to ensure
clean shutdown, a SIGTERM handler is introduced that cancels the main
task of the IVR.

Also adapt the documentation to inform the user that the yate connection
is gone when the hangup handlers are called.

This fixes #9.