Improve callgen speed
When the sound directory is large, iterating over all items
appears to be really slow. For this reason, we change the
behavior from searching all subdirs to explicitly mention
search directories in the commandline. Only these paths
will be searched for sound files. No subdirectories anymore.

This eliminates the need for iteration by only testing if the
searched file exists.